requirements.txt
Django==2.2.4
djangorestframework==3.10.2
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
